requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.12.0
ephem>=4.1.4
skyfield>=1.45.0
astropy>=5.2.2
numpy>=1.24.0
requests>=2.28.0  # For potential future API calls
python-dateutil>=2.8.2  # For date handling
jplephem>=2.18  # For JPL ephemerides